requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
opencv-python>=4.6.0
pillow>=9.3.0
numpy>=1.23.5